+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.18.0] - 2025-09-16
 
 This is unfortunately going to break all `endpoint::list` calls due to backend changes. The good
 news is that you can get more or less the old behavior by adding:
@@ -370,6 +370,7 @@ Disabled `reqwest`'s default features and enabled `rustls-tls`.
 
 Initial release.
 
+[0.18.0]: https://crates.io/crates/resend-rs/0.18.0
 [0.17.1]: https://crates.io/crates/resend-rs/0.17.1
 [0.17.0]: https://crates.io/crates/resend-rs/0.17.0
 [0.16.1]: https://crates.io/crates/resend-rs/0.16.1